requirements.txt
aniso8601==2.0.0
astroid==1.6.1
click==6.7
Flask==0.12.2
Flask-JWT-Simple==0.0.3
Flask-PyMongo==0.5.1
Flask-RESTful==0.3.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
paho-mqtt==1.3.1
PyJWT==1.5.3
pylint==1.8.2
pymongo==3.6.0
python-dotenv==0.7.1
pytz==2018.3
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
